requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
python-dotenv
Flask-Login
Flask-WTF